[SYCL][NFC] Refactor RT unit tests
Merged get_native_interop and spec_constants tests into SYCL2020API test suite,
intorduced a header with common PI API redefinitions.
