+++ sycl/unittests/CMakeLists.txt
@@ -10,12 +10,11 @@ endforeach()
 include(AddSYCLUnitTest)
 
 add_subdirectory(allowlist)
-add_subdirectory(get_native_interop)
 add_subdirectory(misc)
 add_subdirectory(pi)
 add_subdirectory(kernel-and-program)
 add_subdirectory(queue)
 add_subdirectory(scheduler)
-add_subdirectory(spec_constants)
+add_subdirectory(SYCL2020API)
 add_subdirectory(thread_safety)
 add_subdirectory(program_manager)
