+++ sycl/unittests/SYCL2020API/CMakeLists.txt
@@ -2,7 +2,8 @@ set(CMAKE_CXX_EXTENSIONS OFF)
 
 # Enable exception handling for these unit tests
 set(LLVM_REQUIRES_EH 1)
-add_sycl_unittest(SpecConstantsTests OBJECT
-  DefaultValues.cpp
+add_sycl_unittest(SYCL2020APITests OBJECT
+  GetNative.cpp
+  SpecConstDefaultValues.cpp
 )
