+++ sycl/unittests/SYCL2020API/GetNative.cpp
@@ -1,6 +1,4 @@
-//==----------- test_get_native.cpp --- get_native interop unit test only for
-// opencl
-//-------------==//
+//==----------- GetNative.cpp ---  interop unit test only for opencl -------==//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -13,8 +11,11 @@
 #include <CL/sycl.hpp>
 #include <CL/sycl/backend/opencl.hpp>
 #include <detail/context_impl.hpp>
-#include <gtest/gtest.h>
+
 #include <helpers/PiMock.hpp>
+#include <helpers/CommonRedefinitions.hpp>
+
+#include <gtest/gtest.h>
 
 #include <iostream>
 #include <memory>
@@ -48,27 +49,6 @@ static pi_result redefinedEventRetain(pi_event c) {
   return PI_SUCCESS;
 }
 
-static pi_result redefinedProgramCreateWithSource(pi_context context,
-                                                  pi_uint32 count,
-                                                  const char **strings,
-                                                  const size_t *lengths,
-                                                  pi_program *ret_program) {
-  return PI_SUCCESS;
-}
-
-static pi_result
-redefinedProgramBuild(pi_program program, pi_uint32 num_devices,
-                      const pi_device *device_list, const char *options,
-                      void (*pfn_notify)(pi_program program, void *user_data),
-                      void *user_data) {
-  return PI_SUCCESS;
-}
-
-pi_result redefinedEventsWait(pi_uint32 num_events,
-                              const pi_event *event_list) {
-  return PI_SUCCESS;
-}
-
 pi_result redefinedEventGetInfo(pi_event event, pi_event_info param_name,
                                 size_t param_value_size, void *param_value,
                                 size_t *param_value_size_ret) {
@@ -81,9 +61,7 @@ pi_result redefinedEventGetInfo(pi_event event, pi_event_info param_name,
   return PI_SUCCESS;
 }
 
-pi_result redefinedEventRelease(pi_event event) { return PI_SUCCESS; }
-
-TEST(GetNativeTest, GetNativeHandle) {
+TEST(GetNative, GetNativeHandle) {
   platform Plt{default_selector()};
   if (Plt.get_backend() != backend::opencl) {
     std::cout << "Test is created for opencl only" << std::endl;
@@ -97,14 +75,9 @@ TEST(GetNativeTest, GetNativeHandle) {
   TestCounter = 0;
 
   unittest::PiMock Mock{Plt};
-  Mock.redefine<detail::PiApiKind::piclProgramCreateWithSource>(
-      redefinedProgramCreateWithSource);
-  Mock.redefine<detail::PiApiKind::piProgramBuild>(redefinedProgramBuild);
+  setupDefaultMockAPIs(Mock);
 
-  Mock.redefine<detail::PiApiKind::piEventsWait>(redefinedEventsWait);
   Mock.redefine<detail::PiApiKind::piEventGetInfo>(redefinedEventGetInfo);
-  Mock.redefine<detail::PiApiKind::piEventRelease>(redefinedEventRelease);
-
   Mock.redefine<detail::PiApiKind::piContextRetain>(redefinedContextRetain);
   Mock.redefine<detail::PiApiKind::piQueueRetain>(redefinedQueueRetain);
   Mock.redefine<detail::PiApiKind::piDeviceRetain>(redefinedDeviceRetain);
