--- sycl/unittests/get_native_interop/CMakeLists.txt
@@ -1,3 +1,0 @@
-add_sycl_unittest(GetNativeTests OBJECT
-  test_get_native.cpp
-)
