+++ sycl/unittests/helpers/PiImage.hpp
@@ -283,7 +283,7 @@ std::enable_if_t<Idx == sizeof...(Ts)> iterate_tuple(Func &F,
 }
 template <typename Func, uint32_t Idx = 0, typename... Ts>
     std::enable_if_t <
-    Idx<sizeof...(Ts)> iterate_tuple(Func &F, std::tuple<Ts...> &Tuple) {
+    Idx<sizeof...(Ts)> inline iterate_tuple(Func &F, std::tuple<Ts...> &Tuple) {
   const auto &Value = std::get<Idx>(Tuple);
   const char *Begin = reinterpret_cast<const char *>(&Value);
   const char *End = Begin + sizeof(Value);
@@ -301,10 +301,11 @@ template <typename Func, uint32_t Idx = 0, typename... Ts>
 /// \param Offsets is a list of offsets inside composite spec constant.
 /// \param DefaultValues is a tuple of default values for composite spec const.
 template <typename... T>
-PiProperty makeSpecConstant(std::vector<char> &ValData, const std::string &Name,
-                            std::initializer_list<uint32_t> IDs,
-                            std::initializer_list<uint32_t> Offsets,
-                            std::tuple<T...> DefaultValues) {
+inline PiProperty makeSpecConstant(std::vector<char> &ValData,
+                                   const std::string &Name,
+                                   std::initializer_list<uint32_t> IDs,
+                                   std::initializer_list<uint32_t> Offsets,
+                                   std::tuple<T...> DefaultValues) {
   const size_t PropByteArraySize = sizeof...(T) * sizeof(uint32_t) * 3;
   std::vector<char> DescData;
   DescData.resize(8 + PropByteArraySize);
@@ -360,8 +361,8 @@ PiProperty makeSpecConstant(std::vector<char> &ValData, const std::string &Name,
 /// Utility function to add specialization constants to property set.
 ///
 /// This function overrides the default spec constant values.
-void addSpecConstants(PiArray<PiProperty> SpecConstants,
-                      std::vector<char> ValData, PiPropertySet &Props) {
+inline void addSpecConstants(PiArray<PiProperty> SpecConstants,
+                             std::vector<char> ValData, PiPropertySet &Props) {
   Props.insert(__SYCL_PI_PROPERTY_SET_SPEC_CONST_MAP, std::move(SpecConstants));
 
   PiProperty Prop{"all", std::move(ValData), PI_PROPERTY_TYPE_BYTE_ARRAY};
@@ -373,7 +374,7 @@ void addSpecConstants(PiArray<PiProperty> SpecConstants,
 }
 
 /// Utility function to add ESIMD kernel flag to property set.
-void addESIMDFlag(PiPropertySet &Props) {
+inline void addESIMDFlag(PiPropertySet &Props) {
   std::vector<char> ValData(sizeof(uint32_t));
   ValData[0] = 1;
   PiProperty Prop{"isEsimdImage", ValData, PI_PROPERTY_TYPE_UINT32};
@@ -384,7 +385,7 @@ void addESIMDFlag(PiPropertySet &Props) {
 }
 
 /// Utility function to generate offload entries for kernels without arguments.
-PiArray<PiOffloadEntry>
+inline PiArray<PiOffloadEntry>
 makeEmptyKernels(std::initializer_list<std::string> KernelNames) {
   PiArray<PiOffloadEntry> Entries;
